refactor: command_handler をマップ化し register.ts を分割
command_handler.ts の if-else 分岐を Record ベースのディスパッチ
マップへ置換（vclock のガード・実行順・エラー時挙動は維持）。
register.ts のコマンド定義を機能別モジュール（register/admin_commands,
recruit_commands, util_commands, vc_commands）へ分割し、
commandDefinitions の内容・registerSlashCommands のロジックは
変更しない。
